Refactor: Restructure project into a package and separate concerns into modules.
- Moved all Python source files into a `market_analysis` package.
- Updated `pyproject.toml` to reflect the new package structure.
- Created a `__main__.py` for a clean entry point.
- Refactored the monolithic `position_risk_manager.py` into smaller, more focused modules for config, utils, confidence scoring, and reporting.
- Updated all imports to be relative within the package.
- Updated and ran tests to ensure all functionality is preserved.
