+++ market_analysis/atr_sl_gpt.py
@@ -148,77 +148,3 @@ def fetch_ohlcv_df(exchange: ccxt.Exchange, symbol: str, timeframe: str, limit: 
 
 # ---------- CLI ----------
 
-def main():
-    p = argparse.ArgumentParser(description="ATR-based SL/TP sizing and trailing stop")
-    p.add_argument('--exchange', type=str, required=True, help="bybit, binance, okx, etc")
-    p.add_argument('--symbol', type=str, required=True, help="e.g., ETH/USDT:USDT (bybit) or ETH/USDT (binance)")
-    p.add_argument('--timeframe', type=str, default="4h")
-    p.add_argument('--atr', type=int, default=20)
-    p.add_argument('--k', type=float, default=1.4, help="SL multiplier")
-    p.add_argument('--m', type=float, default=2.8, help="TP multiplier")
-    p.add_argument('--entry', type=float, required=True)
-    p.add_argument('--side', type=str, choices=['long','short'], required=True)
-    p.add_argument('--account-risk', type=float, default=100.0, help="Risk per trade in USDT")
-    p.add_argument('--leverage', type=float, default=5.0)
-    p.add_argument('--trail-mult', type=float, default=1.2, help="ATR trailing stop multiplier")
-    p.add_argument('--sandbox', action='store_true', help="Use exchange sandbox if available")
-    args = p.parse_args()
-
-    ex = load_exchange(args.exchange, sandbox=args.sandbox)
-
-    df = fetch_ohlcv_df(ex, args.symbol, args.timeframe, limit=max(100, args.atr*3))
-    df['atr'] = atr_wilder(df, n=args.atr)
-    df['atr_ratio'] = df['atr'] / df['atr'].rolling(args.atr).mean()
-    latest = df.iloc[-1]
-    atr_val = float(latest['atr'])
-    close = float(latest['close'])
-
-    levels = compute_levels(entry=args.entry, atr=atr_val, k=args.k, m=args.m, side=args.side)
-    qty = position_size_usdt(account_risk=args.account_risk, sl_distance=levels.sl_distance)
-
-    liq = approx_liq_price(entry=args.entry, side=args.side, leverage=args.leverage)
-    buffer_ratio = check_liq_buffer(entry=args.entry, sl=levels.sl_price, liq=liq, side=args.side)
-
-    # trailing stop proposal based on last bar close
-    trail = atr_trailing_stop(price=close, atr=atr_val, trail_mult=args.trail_mult, side=args.side, last_trail=None)
-
-    out = {
-        "symbol": args.symbol,
-        "timeframe": args.timeframe,
-        "atr_period": args.atr,
-        "atr_value": round(atr_val, 6),
-        "atr_ratio": round(float(latest['atr_ratio']), 4) if not math.isnan(latest['atr_ratio']) else None,
-        "entry": args.entry,
-        "side": args.side,
-        "k_sl_mult": args.k,
-        "m_tp_mult": args.m,
-        "sl_price": round(levels.sl_price, 6),
-        "tp_price": round(levels.tp_price, 6),
-        "sl_distance": round(levels.sl_distance, 6),
-        "tp_distance": round(levels.tp_distance, 6),
-        "risk_reward": round(levels.rr, 3),
-        "qty_usdt_linear": round(qty, 6),
-        "approx_liq_price": round(liq, 6),
-        "liq_to_sl_buffer_ratio": round(buffer_ratio, 3),
-        "trail_stop_proposal": round(trail, 6),
-    }
-
-    print(json.dumps(out, indent=2))
-
-    # --- Example (commented) of order placement via CCXT ---
-    # ex.load_markets()
-    # market = ex.market(args.symbol)
-    # ex.set_leverage(args.leverage, args.symbol) if hasattr(ex, 'set_leverage') else None
-    # if args.side == 'long':
-    #     # market order open long
-    #     ex.create_order(args.symbol, 'market', 'buy', qty)
-    #     # place SL/TP
-    #     ex.create_order(args.symbol, 'stop_market', 'sell', qty, None, {'stopPrice': levels.sl_price})
-    #     ex.create_order(args.symbol, 'take_profit_market', 'sell', qty, None, {'stopPrice': levels.tp_price})
-    # else:
-    #     ex.create_order(args.symbol, 'market', 'sell', qty)
-    #     ex.create_order(args.symbol, 'stop_market', 'buy', qty, None, {'stopPrice': levels.sl_price})
-    #     ex.create_order(args.symbol, 'take_profit_market', 'buy', qty, None, {'stopPrice': levels.tp_price})
-
-if __name__ == "__main__":
-    main()
