+++ market_analysis/garch_vol_triggers.py
@@ -9,25 +9,7 @@ import numpy as np
 import pandas as pd
 import ccxt
 
-# --- New config loader (optional best-effort) ---
-try:
-    import tomllib  # Python 3.11+
-except Exception:
-    tomllib = None
-
-def load_settings(path: str = "settings.toml") -> dict:
-    if not tomllib:
-        return {}
-    try:
-        with open(path, "rb") as f:
-            return tomllib.load(f)
-    except Exception:
-        # fallback to example if present
-        try:
-            with open("settings.example.toml", "rb") as f:
-                return tomllib.load(f)
-        except Exception:
-            return {}
+from .config import settings
 
 # ---------- CCXT Bybit Data Fetcher ----------
 
@@ -374,154 +356,3 @@ def compute_trailing_stop(entry: float, direction: str, atr: float, cfg: dict, r
     else:
         return entry + atr * trail_mult
 
-# ---------- Enhanced example with multiple symbols ----------
-
-def analyze_multiple_symbols_bybit(exchange: ccxt.Exchange, symbols=["BTC/USDT", "ETH/USDT"], timeframe="4h", days_back=120):
-    """
-    Analyze multiple symbols from Bybit with volatility forecasts and SL/TP calculations.
-    """
-    results = {}
-    
-    for symbol in symbols:
-        print(f"\n=== Analyzing {symbol} ===")
-        
-        # Get market info
-        market_info = get_bybit_market_info(exchange, symbol)
-        tick_size = market_info['tick_size'] if market_info else None
-        
-        # Get data
-        df = get_klines_bybit(exchange, symbol, timeframe,
-                             since=dt.datetime.now(dt.timezone.utc) - dt.timedelta(days=days_back))
-        
-        if df.empty:
-            print(f"No data for {symbol}")
-            continue
-            
-        # Compute ATR
-        df["ATR20"] = compute_atr(df, period=20)
-        
-        # HAR-RV analysis
-        H_hours = 4
-        try:
-            sigma_ann_har, sigma_H_har = sigma_ann_and_sigma_H_from_har(df["close"], interval=timeframe, horizon_hours=H_hours)
-            print(f"[HAR] sigma_ann={sigma_ann_har:.2%}, sigma_H({H_hours}h)={sigma_H_har:.2%}")
-        except Exception as e:
-            print(f"HAR failed: {e}")
-            sigma_ann_har, sigma_H_har = None, None
-        
-        # GARCH analysis
-        try:
-            sigma_ann_garch, sigma_H_garch, garch_res = garch_sigma_ann_and_sigma_H(df["close"], interval=timeframe, horizon_hours=H_hours)
-            print(f"[GARCH] sigma_ann={sigma_ann_garch:.2%}, sigma_H({H_hours}h)={sigma_H_garch:.2%}")
-            # Validate
-            rets = np.log(df["close"]).diff().dropna().values
-            issues = validate_garch_result(rets, garch_res, sigma_ann_garch, H_hours)
-            if issues:
-                print("GARCH checks:", "; ".join(issues))
-        except Exception as e:
-            print(f"GARCH failed: {e}")
-            sigma_ann_garch, sigma_H_garch = None, None
-        
-        # Get live price for current analysis
-        live_price = get_live_price_bybit(exchange, symbol)
-        if live_price is None:
-            print(f"Warning: Could not get live price for {symbol}, using last close price")
-            entry_price = float(df["close"].iloc[-1])
-        else:
-            entry_price = live_price
-            print(f"Live price for {symbol}: ${entry_price:.2f}")
-        
-        # Blended sigma using ATR abs
-        atr = float(df["ATR20"].iloc[-1])
-        cfg = load_settings()
-        sigmaH_blend_abs = blended_sigma_h(sigma_ann_garch, sigma_ann_har, atr_abs=atr, price=entry_price, cfg=cfg)
-        sigmaH_blend_frac = sigmaH_blend_abs / entry_price
-        
-        risk_R = 200.0
-        
-        strategies = {}
-        strategies["VOL_BLEND"] = sl_tp_and_size(entry_price, sigma_H=sigmaH_blend_frac, k=1.2, m=2.2, side="long", R=risk_R, tick_size=tick_size)
-        
-        results[symbol] = {
-            'market_info': market_info,
-            'entry_price': entry_price,
-            'strategies': strategies,
-            'sigma_ann_har': sigma_ann_har if sigma_ann_har else None,
-            'sigma_ann_garch': sigma_ann_garch if sigma_ann_garch else None,
-            'atr': atr,
-            'sigmaH_blend_abs': sigmaH_blend_abs
-        }
-        
-        # Print strategy results
-        for strategy_name, params in strategies.items():
-            print(f"[{strategy_name}] SL=${params['SL']:.2f}, TP=${params['TP']:.2f}, Size={params['Q']:.4f}")
-    
-    return results
-
-# ---------- Example usage ----------
-
-if __name__ == "__main__":
-    from dotenv import load_dotenv
-    load_dotenv()
-
-    # Set to True to use Bybit testnet
-    USE_SANDBOX = False
-    
-    # For standalone execution, create an exchange instance
-    exchange = ccxt.bybit({
-        'sandbox': USE_SANDBOX,
-        'enableRateLimit': True,
-        'apiKey': os.getenv('BYBIT_API_KEY'),
-        'secret': os.getenv('BYBIT_API_SECRET'),
-    })
-
-    # Single symbol analysis
-    print("=== Single Symbol Analysis (BTC/USDT) ===")
-    df = get_klines_bybit(exchange, "BTC/USDT", "4h",
-                         since=dt.datetime.now(dt.timezone.utc) - dt.timedelta(days=120))
-    
-    if not df.empty:
-        print(f"Downloaded {len(df)} candles, latest close: ${df['close'].iloc[-1]:.2f}")
-        
-        # Get market info
-        market_info = get_bybit_market_info(exchange, "BTC/USDT")
-        if market_info:
-            print(f"Tick size: {market_info['tick_size']}, Min quantity: {market_info['min_quantity']}")
-        
-        # Run analysis
-        df["ATR20"] = compute_atr(df, period=20)
-        H_hours = 4
-        sigma_ann_har, sigma_H_har = sigma_ann_and_sigma_H_from_har(df["close"], interval="4h", horizon_hours=H_hours)
-        
-        # Get live price for current analysis
-        live_price = get_live_price_bybit(exchange, "BTC/USDT")
-        if live_price is None:
-            print(f"Warning: Could not get live price, using last close price")
-            entry_price = float(df["close"].iloc[-1])
-        else:
-            entry_price = live_price
-            print(f"Live BTC price: ${entry_price:.2f}")
-        
-        # Use blended vol
-        cfg = load_settings()
-        atr_val = float(df["ATR20"].iloc[-1])
-        sigma_ann_garch, sigma_H_garch, garch_res = garch_sigma_ann_and_sigma_H(df["close"], interval="4h", horizon_hours=H_hours)
-        sigmaH_blend_abs = blended_sigma_h(sigma_ann_garch, sigma_ann_har, atr_abs=atr_val, price=entry_price, cfg=cfg)
-        params = sl_tp_and_size(entry_price, sigma_H=(sigmaH_blend_abs/entry_price), k=1.2, m=2.2, side="long", R=200.0, 
-                               tick_size=market_info['tick_size'] if market_info else None)
-        print(f"Dynamic SL/TP (blended): {params}")
-    
-    # Multiple symbols analysis
-    print("\n=== Multiple Symbols Analysis ===")
-    symbols = ["BTC/USDT", "ETH/USDT", "SOL/USDT"]
-    results = analyze_multiple_symbols_bybit(exchange, symbols, timeframe="4h", days_back=90)
-    
-    # Summary table
-    print("\n=== Summary Table ===")
-    print(f"{'Symbol':<12} {'Price':<10} {'HAR σ_ann':<10} {'ATR':<8} {'Blend SL':<8} {'Blend TP':<8}")
-    print("-" * 70)
-    for symbol, data in results.items():
-        har_sigma = f"{data['sigma_ann_har']:.1%}" if data['sigma_ann_har'] else "N/A"
-        blend_sl = f"${data['strategies']['VOL_BLEND']['SL']:.0f}" if 'VOL_BLEND' in data['strategies'] else "N/A"
-        blend_tp = f"${data['strategies']['VOL_BLEND']['TP']:.0f}" if 'VOL_BLEND' in data['strategies'] else "N/A"
-        print(f"{symbol:<12} ${data['entry_price']:<9.0f} {har_sigma:<10} {data['atr']:<7.0f} {blend_sl:<8} {blend_tp:<8}")
